+++ front-page.php
@@ -17,18 +17,19 @@ get_header();
 ?>
 	
 	<!-- Content Container -->
-	<div class="large-9 columns">
+	<div class="large-12 columns">
 		<!-- Slider and Main Container -->
 		<div class="row">
 			<!-- Slider -->
-			<div class="large-12 columns">
+			<div class="large-9 columns">
 				<?php foundation_orbit(); ?>
 			</div>
-
 			<!-- End Slider -->
-
+			<?php make_frontpage_widget_primary_one(); ?>
+		</div>
+		<div class="row">
 			<!-- Main Content -->
-			<div class="large-12 columns" role="main">
+			<div class="large-8 columns" role="main">
 				<?php
 				if( have_posts() ):
 					while( have_posts() ) : the_post();
@@ -41,12 +42,12 @@ get_header();
 					<?php
 				endif;
 				?>
-
 			</div>
 			<!-- End Main Content -->
+			<?php make_frontpage_widget_primary_two(); ?>
 		</div>
 		<!-- End Slider and Main Container -->
 	</div>
 	<!-- End Content Container -->
-<?php get_sidebar(); ?>
+<?php //get_sidebar(); ?>
 <?php get_footer(); ?>