+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 ### HEAD
-* Initial commit
+
+July 8, 2013
+* Made frontpage
+* added frontpage widget areas
+
+July 7, 2013
+* Initial commit of theme files